ui improved

+++ src/services/adminNotice.service.ts
@@ -1,4 +1,5 @@
 import api from "@/lib/axios";
+import Swal from "sweetalert2";
 
 export const getAdminNotices = async (params: {
     page: number;
@@ -18,7 +19,24 @@ export const getNoticeCounts = async () => {
 };
 
 export const softDeleteNotice = async (id: string) => {
-    return api.delete(`/notices/${id}`);
+    Swal.fire({
+        title: "Are you sure?",
+        text: "You will able to revert this from Trash!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, move to trash."
+    }).then((result) => {
+        if (result.isConfirmed) {
+            Swal.fire({
+                title: "Deleted!",
+                text: "Notice has been moved to Trash.",
+                icon: "success"
+            });
+            return api.delete(`/notices/${id}`);
+        }
+    });
 };
 
 // single restore
@@ -34,8 +52,24 @@ export const bulkRestoreNotice = async (ids: string[]) => {
 };
 
 export const permanentDeleteNotices = async (ids: string[]) => {
-    // return api.post(`/notices/permanent`, { ids });
-    return api.delete(`/notices/permanent`, {
-        data: { ids }
+    Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+        if (result.isConfirmed) {
+            Swal.fire({
+                title: "Deleted!",
+                text: "Notice has been deleted.",
+                icon: "success"
+            });
+            return api.delete(`/notices/permanent`, {
+                data: { ids }
+            });
+        }
     });
 };
